requirements.txt
# Core dependencies
pyusb>=1.3.1
Pillow>=9.0.0
mss>=6.0.0

# System monitoring
psutil>=5.8.0

# NVIDIA GPU monitoring
pynvml>=11.0.0

# Development and testing (optional)
# fonttools>=4.0.0


